Clean up AstConv
